+++ src/persistence/puzzleRepository.js
@@ -1,4 +1,4 @@
-import { getValue, setValue } from './jsonFileStore.js';
+import { getValue, setValue } from './userDataFile.js';
 
 /**
  * This repository serves as an abstraction between how the data is stored
