+++ src/persistence/rateLimitRepository.js
@@ -1,5 +1,5 @@
 import { isDate, isValid, parseISO } from 'date-fns';
-import { getValue, setValue } from './jsonFileStore.js';
+import { getValue, setValue } from './userDataFile.js';
 
 /**
  * The key to the data in the json file.
