+++ tests/persistence/jsonFileStore.test.js
@@ -25,13 +25,13 @@ jest.unstable_mockModule('../../src/persistence/cachedValue.js', () => ({
 // import after setting up the mock so the modules import the mocked version
 // const { CachedValue } = await import('../src/persistence/cachedValue.js');
 const { loadFileContents, saveFile, fileExists } = await import('../../src/persistence/io.js');
-const { getValue, setValue, userDataFileExists } = await import('../../src/persistence/jsonFileStore.js');
+const { getValue, setValue, userDataFileExists } = await import('../../src/persistence/userDataFile.js');
 
 beforeEach(() => {
   jest.resetAllMocks();
 });
 
-describe('jsonFileStore', () => {
+describe('userDataFile', () => {
   describe('getValue()', () => {
     describe('cache is empty', () => {
       // mock empty cache for each test in this block.
