+++ tests/persistence/rateLimitRepository.test.js
@@ -3,13 +3,13 @@ import {
 } from '@jest/globals';
 
 // setup mocks.
-jest.unstable_mockModule('../../src/persistence/jsonFileStore.js', () => ({
+jest.unstable_mockModule('../../src/persistence/userDataFile.js', () => ({
   getValue: jest.fn(),
   setValue: jest.fn(),
 }));
 
 // import after setting up the mock so the modules import the mocked version
-const { getValue, setValue } = await import('../../src/persistence/jsonFileStore.js');
+const { getValue, setValue } = await import('../../src/persistence/userDataFile.js');
 const { getRateLimit, setRateLimit } = await import('../../src/persistence/rateLimitRepository.js');
 
 afterEach(() => {
